have complete non-Arduino test program, fixed a couple bugs found in the process, and added documentation about the 1s-complement bit format that seems to be in use

+++ DHT.h
@@ -100,7 +100,7 @@ class DHT {
 
 		uint8_t pin_, type_;
 
-		uint8_t minSampleDelayMillis_;
+		uint16_t minSampleDelayMillis_;
 		uint8_t data_[6];
 		unsigned long lastReadTime_;
 		boolean firstReading_;
